Updates to work on live-1 cluster
We need to make some changes to the Dockerfile in order for the
application to work in the new K8s live-1 cluster.

In Live-1, the Cloud Platform team introduced Pod Security Policies,
to tighten the security on this production cluster.

The main impact of this restricted policy is that it prevents
pods/containers from running as the root user.