Ignore unused image assets requests
This seems to be the result of some kind of bot or crawler, but every 3 days, our logs get full of 404 due to requests to image assets we don't use in this project.

This is usually not a problem, but we've setup some prometheus alerts that gets triggered easily when this huge amount of 404 errors happen in a short amount of time (seconds).

With this small workaround we are going to "ignore" (returning an early 200 response) these requests to these specific assets. Any other asset is not affected.