Bump dependencies
Bump outdated dependencies in an attempt to reduce the number of security warnings on github.

Most of the warnings come from the legacy `pflr-express-kit` so probably not much to do unless we bump the dependencies in that package.

These warnings however does not impact directly the application as it is a compiled static website.