Mitigate a redirection vulnerability
Ticket: https://trello.com/c/9VDqIYdW

This is a bit dirty but will do the work until maybe a more solid solution is found (deep in the framework we are using which is the seed of what is now Form Builder).

Essentially, we rewrite the url by squishing 2 or more slashes `/` into just one, thus eliminating the redirection issue. It will just return a 404 not found page.