Add google optimize script tag
Ticket: https://trello.com/c/kOng50jf

This is for now a proof of concept. In essence it works in a very similar way to google tag manager.
Nothing visible yet, WIP.

The installation recommends the script to be added at the top of the `<HEAD>` tag. Thus, using the stylesheet block, that gets rendered inside the HEAD.